requirements.txt
appdirs==1.4.3
asgiref==3.2.3
backcall==0.1.0
bcrypt==3.1.7
bleach==1.4.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudinary==1.11.0
colorama==0.4.3
confusable-homoglyphs==3.2.0
coverage==5.0.3
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==2.1.5
django-allauth==0.41.0
django-countries==6.0
django-coverage==1.2.4
django-crispy-forms==1.9.0
django-hvad==1.8.0
django-libs==2.0.2
django-registration==2.4.1
django-restframework==0.0.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-registration==0.0.2
entrypoints==0.3
Fabric3==1.14.post1
Faker==0.9.1
filelock==3.0.12
flake8==3.7.9
gunicorn==19.6.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
ipdb==0.13.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Markdown==3.2.1
mccabe==0.6.1
mixer==6.1.3
mock==4.0.1
oauthlib==3.1.0
packaging==20.1
paramiko==2.7.1
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.3
psycopg2==2.7.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
python-dateutil==2.8.1
python-decouple==3.0
python3-openid==3.1.0
pytz==2019.3
reportlab==3.5.34
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
stripe==2.43.0
text-unidecode==1.2
toml==0.10.0
tox==3.14.5
traitlets==4.3.3
Unipath==1.0
urllib3==1.25.8
virtualenv==20.0.7
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==3.2
xhtml2pdf==0.2.4
zipp==3.0.0
